move deploy preferences widgets to a separate class to allow reuse
