+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy.ui/src/com/google/cloud/tools/eclipse/appengine/deploy/ui/messages.properties
@@ -4,7 +4,7 @@ google.preferences=Google Cloud Platform-related settings
 # Deploy settings
 auto.promote=Promote the deployed version to receive all traffic
 bucket.name=Bucket:
-cannot.open.browser=Cannot open browser: {0}
+cannot.open.browser=Cannot open browser
 deploy.login.dialog.message=Browser opened to authorize deployment.
 deploy.manual.link=To manually promote a version use the <a href="{0}">Google Cloud Console.</a>
 deploy.preferences.save.error.title=Could not save preferences
